msm: kgsl: Don't silently ignore events that have a retired timestamp
After all the current events are processed in the main loop, the next
pending event is marked to fire an interrupt to start the cycle anew.
When the event is marked, we check the timestamp again to make sure that
we are always caught up with the GPU.  If the timestamp for the next
event happens to be retired when we get to this point, we silently leave
the function.

If there are no other subsequent events or waittimestamp requests, then
it is possible that another interrupt might never be fired for that context
and then the event gets missed.  Return a 1 in this case all the way back
up the stack and make the event code keep processing events until we are
positive that none are past.

CRs-fixed: 447655
Signed-off-by: Jordan Crouse <jcrouse@codeaurora.org>
(cherry picked from commit f64ee0494159e61dbac2868162878d5e9a618587)

Change-Id: I0c83d8370098357fdc28d8b995a132593bab2ebd
Signed-off-by: Srivalli Oguri <oguri@codeaurora.org>