msm: kgsl: Don't change the sglen of the memdesc if allocation fails
If the page_alloc loop fails, we changed the length of the scatterlist
for the memdesc to the last good page so we could have an efficient
loop through kgsl_sharedmem_free.  Unfortunately, we were also using
the sglen to decide between using kmem or vmalloc memory for the sg
list in the first place. There is a chance that one could start out
with a large sg list from vmalloc, but if it failed early, then we
could truncate the size and attempt to kfree it.

Zero the stlist at init time and use a length == 0 as an indicator
that the list might be truncated.
CRs-Fixed: 443441

Signed-off-by: Dhivya Subramanian <dthiru@codeaurora.org>
(cherry picked from commit 24866eefe7db32ab8f32ba69518f65f1c74c66c2)

Change-Id: Id1d2052ae034db78f42cc5fe6bb742ae8130f391
Signed-off-by: Srivalli Oguri <oguri@codeaurora.org>